% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall R code for my package(Dna).R,
%   R/functions.R
\name{calculate_gc_content}
\alias{calculate_gc_content}
\title{Function to calculate GC content}
\usage{
calculate_gc_content(sequence)

calculate_gc_content(sequence)
}
\arguments{
\item{sequence}{A character vector representing a DNA sequence.}
}
\value{
A numeric value representing the GC content of the input DNA sequence

A numeric value representing the GC content of the input DNA sequence
}
\description{
Function to calculate GC content

Title Calculate GC Content
}
