% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall R code for my package(Dna).R,
%   R/functions.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Function to convert DNA to RNA}
\usage{
dna_to_rna(dna_sequence)

dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character vector representing a DNA sequence.}
}
\value{
A character vector representing the RNA sequence derived from the input DNA sequence

A character vector representing the RNA sequence derived from the input DNA sequence
}
\description{
Function to convert DNA to RNA

Title Transcribe DNA to RNA
}
