% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall R code for my package(Dna).R, R/S3
%   class.R
\name{FastaSeq}
\alias{FastaSeq}
\alias{FastaSeq-class}
\title{FASTA Sequence Class}
\usage{
FastaSeq(header, sequence)

FastaSeq(header, sequence)
}
\description{
Create S3 class for FASTA sequences

Create S3 class for FASTA sequences
}
\section{Slots}{

\describe{
\item{\code{header}}{The header of the FASTA sequence.}

\item{\code{sequence}}{The DNA sequence.}

\item{\code{header}}{The header of the FASTA sequence.}

\item{\code{sequence}}{The DNA sequence.}
}}

