% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall R code for my package(Dna).R,
%   R/functions.R
\name{rna_to_protein}
\alias{rna_to_protein}
\title{Function to translate RNA to protein sequence}
\usage{
rna_to_protein(rna_sequence)

rna_to_protein(rna_sequence)
}
\arguments{
\item{rna_sequence}{rna_sequence A string containing an RNA sequence.}
}
\value{
A protein sequence.

A protein sequence.
}
\description{
Function to translate RNA to protein sequence

Title Translate RNA to Protein
}
