% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall R code for my package(Dna).R,
%   R/functions.R
\name{upload_fasta_sequence}
\alias{upload_fasta_sequence}
\title{Upload FASTA sequence from file}
\usage{
upload_fasta_sequence(file.path)

upload_fasta_sequence(file.path)
}
\arguments{
\item{file.path}{The file path to the FASTA file containing the sequence data}
}
\value{
An object of class 'FastaSeq'.

A list with two elements: 'header' containing the FASTA header and 'sequence' containing the concatenated sequence data.
}
\description{
Upload FASTA sequence from file

Title
}
